# miraculous-kingdom-frontend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "miraculous-kingdom-self-server",
        "description": "",
        "license": { "name": "" },
        "version": "0.1.0"
    },
    "paths": {
        "/api/character/{secret}": {
            "get": {
                "tags": ["api::character_api"],
                "summary": "Endpoint to find all characters that the players is participating in for their specific secret.",
                "description": "Endpoint to find all characters that the players is participating in for their specific secret.\n\n# Example\n\n```\nGET /api/character/{secret}\n```\n\n# Parameters\n\n- `secret`: String set by the user. Should keep the same. For now too lazy to fix 👿\n\n# Responses\n\n- `200 OK`: Found characters with secret_code: {secret}\n\n```json\n{\n\"code\": 200,\n\"message\": \"Found characters with secret_code: {secret}\",\n\"success\": true,\n\"data\": [\n{\n\"char_name\": \"character_name\",\n...\n},\n...\n]\n}\n```\n\n- `500 Internal Server Error`: An internal error occurred.\n\n```json\n{\n\"code\": 500,\n\"success\": false\n\"message\": \"Internal error occurred\",\n\"data\": []\n}\n```\n",
                "operationId": "get_characters",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "path",
                        "description": "String set by the user to get their data",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found characters with secret_code: {secret}",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VecCharDetailedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VecCharDetailedResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/character/{secret}/{pass}": {
            "get": {
                "tags": ["api::character_api"],
                "operationId": "get_character_for_game",
                "parameters": [
                    {
                        "name": "secret",
                        "in": "path",
                        "description": "String set by the user to get their data",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "pass",
                        "in": "path",
                        "description": "String generated by the api for the specific game.",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found characters with : {secret} and {pass}",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharDetialedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharDetialedResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/game": {
            "get": {
                "tags": ["api::game_api"],
                "operationId": "get_games",
                "responses": {
                    "200": {
                        "description": "Game found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GamesInfoDetailedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GamesInfoDetailedResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["api::game_api"],
                "operationId": "start_game",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameCreation"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Found class from database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PassDetailedResponse"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Could not find class from database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PassDetailedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": " Internal error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PassDetailedResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/game/{pass}": {
            "get": {
                "tags": ["api::game_api"],
                "operationId": "get_game",
                "parameters": [
                    {
                        "name": "pass",
                        "in": "path",
                        "description": "Password for entering the game.",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found class from database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInfoDetailedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Could not find class from database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInfoDetailedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": " Internal error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameInfoDetailedResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["api::game_api"],
                "operationId": "add_character",
                "parameters": [
                    {
                        "name": "pass",
                        "in": "path",
                        "description": "Password for entering the game.",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewCharacter"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Added Character to Game",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharAddedDetailedResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharAddedDetailedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Could not find class from database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharAddedDetailedResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": " Internal error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharAddedDetailedResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIError": {
                "type": "object",
                "required": ["status_code", "message"],
                "properties": {
                    "message": { "type": "string" },
                    "status_code": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    }
                }
            },
            "Ability": {
                "type": "object",
                "required": ["ability_name", "ability_desc", "ability_unlock"],
                "properties": {
                    "ability_clock": {
                        "allOf": [{ "$ref": "#/components/schemas/Clock" }],
                        "nullable": true
                    },
                    "ability_desc": { "type": "string" },
                    "ability_name": { "type": "string" },
                    "ability_unlock": {
                        "$ref": "#/components/schemas/MightRequirement"
                    }
                }
            },
            "CharAddedDetailedResponse": {
                "type": "object",
                "required": ["data", "success"],
                "properties": {
                    "data": { "$ref": "#/components/schemas/ClassResponse" },
                    "success": { "$ref": "#/components/schemas/Progress" }
                }
            },
            "CharDetialedResponse": {
                "type": "object",
                "required": ["data", "success"],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/CharacterResponse"
                    },
                    "success": { "$ref": "#/components/schemas/Progress" }
                }
            },
            "CharacterResponse": {
                "type": "object",
                "required": [
                    "secret",
                    "char_name",
                    "char_class",
                    "char_clocks",
                    "char_might",
                    "abilities",
                    "char_state"
                ],
                "properties": {
                    "abilities": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Ability" },
                        "description": "The character's abilities."
                    },
                    "char_class": {
                        "$ref": "#/components/schemas/ClassResponse"
                    },
                    "char_clocks": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Clock" },
                        "description": "The character's clocks."
                    },
                    "char_might": { "$ref": "#/components/schemas/Might" },
                    "char_name": {
                        "type": "string",
                        "description": "The name of the character."
                    },
                    "char_state": {
                        "$ref": "#/components/schemas/CharacterState"
                    },
                    "secret": {
                        "type": "string",
                        "description": "A secret lock to the character."
                    }
                }
            },
            "CharacterState": {
                "type": "string",
                "description": "This enum represents the possible states of a character in a websocket game.\n- Waiting: the character is waiting for its turn to come. - Going: the character is currently taking its turn.\n- Gone: the character has already taken its turn and is no longer in play.",
                "enum": ["Waiting", "Going", "Gone"]
            },
...
            "ClassEnum": {
                "type": "string",
                "enum": [
                    "WarGeneral",
                    "Aficianado",
                    "Researcher",
                    "Cardinal",
                    "SpyMaster",
                    "Diplomat",
                    "Merchant",
                    "Prince"
                ]
            },
            "ClassResponse": {
                "type": "object",
                "required": [
                    "class_enum",
                    "class_desc",
                    "class_perks",
                    "class_abilities",
                    "class_name"
                ],
                "properties": {
                    "class_abilities": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Ability" }
                    },
                    "class_desc": { "type": "string" },
                    "class_enum": { "$ref": "#/components/schemas/ClassEnum" },
                    "class_name": { "type": "string" },
                    "class_perks": { "type": "string" }
                }
            },
            "Clock": {
                "type": "object",
                "required": [
                    "clock_duration",
                    "clock_remaining",
                    "clock_name",
                    "clock_desc",
                    "clock_conf"
                ],
                "properties": {
                    "clock_conf": { "type": "boolean" },
                    "clock_desc": { "type": "string" },
                    "clock_duration": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    },
                    "clock_name": { "type": "string" },
                    "clock_remaining": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    }
                }
            },
            "GameCreation": {
                "type": "object",
                "required": ["game_num_players", "game_name", "game_ruler"],
                "properties": {
                    "game_name": { "type": "string" },
                    "game_num_players": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    },
                    "game_ruler": { "type": "string" }
                }
            },
            "GameInfo": {
                "type": "object",
                "required": [
                    "game_name",
                    "game_ruler",
                    "game_chars",
                    "game_pass"
                ],
                "properties": {
                    "game_chars": {
                        "type": "array",
                        "items": { "type": "string" }
                    },
                    "game_name": { "type": "string" },
                    "game_pass": { "type": "string" },
                    "game_ruler": { "type": "string" }
                }
            },
            "GameInfoDetailedResponse": {
                "type": "object",
                "required": ["data", "success"],
                "properties": {
                    "data": { "$ref": "#/components/schemas/GameInfo" },
                    "success": { "$ref": "#/components/schemas/Progress" }
                }
            },
            "GamesInfoDetailedResponse": {
                "type": "object",
                "required": ["data", "success"],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/GameInfo" }
                    },
                    "success": { "$ref": "#/components/schemas/Progress" }
                }
            },
            "Might": {
                "type": "object",
                "required": [
                    "might_military",
                    "might_culture",
                    "might_religion",
                    "might_science",
                    "might_diplomacy",
                    "might_espionage"
                ],
                "properties": {
                    "might_culture": {
                        "$ref": "#/components/schemas/MightStat"
                    },
                    "might_diplomacy": {
                        "$ref": "#/components/schemas/MightStat"
                    },
                    "might_espionage": {
                        "$ref": "#/components/schemas/MightStat"
                    },
                    "might_military": {
                        "$ref": "#/components/schemas/MightStat"
                    },
                    "might_religion": {
                        "$ref": "#/components/schemas/MightStat"
                    },
                    "might_science": {
                        "$ref": "#/components/schemas/MightStat"
                    }
                }
            },
            "MightEnum": {
                "type": "string",
                "enum": [
                    "Military",
                    "Culture",
                    "Religion",
                    "Science",
                    "Diplomacy",
                    "Espionage"
                ]
            },
            "MightRequirement": {
                "type": "object",
                "required": ["might", "roll_tier", "unlock"],
                "properties": {
                    "might": { "$ref": "#/components/schemas/MightEnum" },
                    "roll_tier": { "$ref": "#/components/schemas/RollTier" },
                    "unlock": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    }
                }
            },
            "MightStat": {
                "type": "object",
                "required": [
                    "stat_enum",
                    "stat_name",
                    "stat_value",
                    "stat_exp",
                    "stat_token"
                ],
                "properties": {
                    "stat_enum": { "$ref": "#/components/schemas/MightEnum" },
                    "stat_exp": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    },
                    "stat_name": { "type": "string" },
                    "stat_token": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0.0
                    },
                    "stat_value": { "type": "integer", "format": "int32" }
                }
            },
            "NewCharacter": {
                "type": "object",
                "description": "A struct representing a new character request from a client.",
                "required": ["secret", "char_name", "char_class", "char_might"],
                "properties": {
                    "char_class": { "$ref": "#/components/schemas/ClassEnum" },
                    "char_might": {
                        "type": "object",
                        "description": "The character's might.",
                        "additionalProperties": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0.0
                        }
                    },
                    "char_name": {
                        "type": "string",
                        "description": "The name of the character."
                    },
                    "secret": {
                        "type": "string",
                        "description": "A secret lock to the character for getting the character."
                    }
                }
            },
            "PassDetailedResponse": {
                "type": "object",
                "required": ["data", "success"],
                "properties": {
                    "data": { "type": "string" },
                    "success": { "$ref": "#/components/schemas/Progress" }
                }
            },
            "Progress": {
                "oneOf": [
                    { "type": "string", "enum": ["Succeeding"] },
                    {
                        "type": "object",
                        "required": ["Failing"],
                        "properties": {
                            "Failing": {
                                "$ref": "#/components/schemas/APIError"
                            }
                        }
                    }
                ]
            },
...
            "VecCharDetailedResponse": {
                "type": "object",
                "required": ["data", "success"],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CharacterResponse"
                        }
                    },
                    "success": { "$ref": "#/components/schemas/Progress" }
                }
            },
...
        }
    },
...
}